wip: initial page setup
