+++ src/ui/components/layout/ResponsivePageLayout/ResponsivePageLayout.types.ts
@@ -4,6 +4,7 @@ interface ResponsivePageLayoutProps {
   header?: ReactNode;
   pageId?: string;
   children?: ReactNode;
+  additionalClassNames?: string;
 }
 
 export type { ResponsivePageLayoutProps };
