--- src/ui/pages/ConnectionCredentialRequest/index.ts
@@ -1,1 +1,0 @@
-export * from "./ConnectionCredentialRequest";
